refactor(image-upload): extract shared image upload draft hook
- centralize file preview, crop completion, upload locking, and reset in useImageUploadDraft
- wire inline and modal image upload UIs through the hook for a single upload flow
- support modal callbacks and parent loading via onCropComplete and isExternallyBusy
